swagger_server/controllers/additional_info_controller.py
import connexion
import six

from swagger_server.models.inline_response200 import InlineResponse200  # noqa: E501
from swagger_server.models.inline_response2001 import InlineResponse2001  # noqa: E501
from swagger_server.models.inline_response2002 import InlineResponse2002  # noqa: E501
from swagger_server.models.inline_response2003 import InlineResponse2003  # noqa: E501
from swagger_server.models.inline_response2004 import InlineResponse2004  # noqa: E501
from swagger_server.models.inline_response2005 import InlineResponse2005  # noqa: E501
from swagger_server.models.inline_response2006 import InlineResponse2006  # noqa: E501
from swagger_server.models.inline_response201 import InlineResponse201  # noqa: E501
from swagger_server.models.users_user_id_body import UsersUserIdBody  # noqa: E501
from swagger_server.models.users_user_id_body1 import UsersUserIdBody1  # noqa: E501
from swagger_server import util


def add_numeric_review_for_content(body, content_id, user_id):  # noqa: E501
    """Add a numeric review for content

    Submit a numeric review for a specific content by a user # noqa: E501

    :param body: The review data
    :type body: dict | bytes
    :param content_id: The ID of the content to review
    :type content_id: int
    :param user_id: The ID of the user submitting the review
    :type user_id: int

    :rtype: InlineResponse201
    """
    if connexion.request.is_json:
        body = UsersUserIdBody1.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def delete_specific_review_for_content_by_user(content_id, user_id):  # noqa: E501
    """Delete a specific review for content by user

    Delete a specific numeric review for a given content and user # noqa: E501

    :param content_id: The ID of the content to delete the review for
    :type content_id: int
    :param user_id: The ID of the user whose review is to be deleted
    :type user_id: int

    :rtype: None
    """
    return 'do some magic!'


def get_content_languages(content_id):  # noqa: E501
    """Get available languages for content

    Retrieve a list of available languages for a specific content # noqa: E501

    :param content_id: The ID of the content to retrieve languages for
    :type content_id: int

    :rtype: InlineResponse2005
    """
    return 'do some magic!'


def get_content_views(content_id):  # noqa: E501
    """Get number of views for content

    Retrieve the number of views for a specific content # noqa: E501

    :param content_id: The ID of the content to retrieve views for
    :type content_id: int

    :rtype: InlineResponse200
    """
    return 'do some magic!'


def get_numeric_review_for_content_by_user(content_id, user_id):  # noqa: E501
    """Get a specific review for content by user

    Retrieve a specific numeric review made by a user for a given content # noqa: E501

    :param content_id: The ID of the content to retrieve the review for
    :type content_id: int
    :param user_id: The ID of the user whose review to retrieve
    :type user_id: int

    :rtype: InlineResponse2002
    """
    return 'do some magic!'


def get_numeric_reviews_by_user(user_id):  # noqa: E501
    """Get reviews by user

    Retrieve all numeric reviews made by a specific user # noqa: E501

    :param user_id: The ID of the user whose reviews to retrieve
    :type user_id: int

    :rtype: InlineResponse2004
    """
    return 'do some magic!'


def get_numeric_reviews_for_content(content_id):  # noqa: E501
    """Get numeric reviews for specific content

    Retrieve all numeric reviews (user ID and rating) made for a specific content # noqa: E501

    :param content_id: The ID of the content to retrieve reviews for
    :type content_id: int

    :rtype: InlineResponse2001
    """
    return 'do some magic!'


def get_recommendations_for_user(user_id, profile_id):  # noqa: E501
    """Get recommendations for a profile

    Retrieve content recommendations based on a user&#x27;s viewing history # noqa: E501

    :param user_id: The ID of the user to retrieve recommendations for
    :type user_id: int
    :param profile_id: The ID of the profile to retrieve recommendations for
    :type profile_id: int

    :rtype: InlineResponse2006
    """
    return 'do some magic!'


def update_specific_review_for_content_by_user(body, content_id, user_id):  # noqa: E501
    """Update a specific review for content by user

    Update an existing numeric review for a given content and user # noqa: E501

    :param body: The updated review data
    :type body: dict | bytes
    :param content_id: The ID of the content to update the review for
    :type content_id: int
    :param user_id: The ID of the user whose review is being updated
    :type user_id: int

    :rtype: InlineResponse2003
    """
    if connexion.request.is_json:
        body = UsersUserIdBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'
